[REF] runbot: move teams stuff
